Move Stitches Logo into `/components`
